borrar marca que no tenga ningun producto asociado

+++ app/Http/Controllers/BrandController.php
@@ -9,7 +9,7 @@ use DataTables;
 class BrandController extends Controller
 {
     //
-    
+
     public function index()
     {
         /*Esta funcion devuelve todas las marcas ordenadas por el nombre
@@ -27,14 +27,14 @@ class BrandController extends Controller
         Parametros : no se le pasa ningun parametro
         Return : una tabla con los atributos de las marcas ordenadas por nombres */
         return DataTables::of(Brand::orderby('brand_name')->get())
-    ->addColumn('accion', function($b){
-      return '';
-  })->make(true);
-}
+        ->addColumn('accion', function($b){
+          return '';
+      })->make(true);
+    }
 
 
-public function store(Request $request)
-{
+    public function store(Request $request)
+    {
     /*Estafuncion Store valida el nombre de la marca , crea una marca y guarda los datos 
     que tiene el request y los guarda en la base de datos
     Parámetros:recibe el parametro request donde tendran los datos de la nueva marca a registrar
@@ -66,15 +66,23 @@ public function destroy($id)
     /*Estafuncion Destroy lo que haces es  buscar la marca con ese id y lo elimina 
     Parametros:recibe el parametro id ,que es el id de la marca a eliminar  
     Return:devuelve si pudo eliminar la marca o no */
-    $brand = Brand::find($id);
-    if ($brand->delete()) {
-        return response()->json([
-            'success' => true
-        ]);
-    } else {
+    $brand = Brand::find($id)->products;
+    if ($brand == null){
+        if ($brand->delete()) {
+            return response()->json([
+                'success' => true
+            ]);
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'La Marca no pudo ser borrada'
+            ], 500);
+        }
+    }
+    else{
         return response()->json([
             'success' => false,
-            'message' => 'La Marca no pudo ser borrada'
+            'message' => 'La Marca no puede ser borrada. tiene un producto asociado'
         ], 500);
     }
 }
